feat: Split config into default (current) and base (without a react dependency)
